requirements.txt
matplotlib
newspaper3k
nltk
pandas
pymongo
scrapy
seaborn